% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crossover}
\alias{crossover}
\title{Chromosomes Crossover and Mutation}
\usage{
crossover(ind1,ind2)
}
\arguments{
\item{ind1,ind2}{a numeric vectors represents individual parent chromosome.}
}
\value{
It returns a matrix with each column represents each offspring from a process of crossover
and mutation.
}
\description{
Make 2 individual parent chromosomes crossover and mute when breeding offsprings.
}
\details{
This function makes two individual parent chromosomes perfrom crossover and mutation when
breeding next generation. Note that the crossover is simply one-point crossover and the mutation is based on mutation\{GARVas\}.
}
\examples{
ind1=init(9)[,1]
ind2=init(9)[,2] #Generate 2 chromosomes
crossover(ind1,ind2)
}

