% Generated by roxygen2 (4.0.2): do not edit by hand
\name{init}
\alias{init}
\title{Generate the 1st generation of the genetic algorithm}
\usage{
init(C)
}
\arguments{
\item{C}{Chromosome length (number of predictors)}
}
\value{
Matrix, [C x 2C] or [C x 50], with each column representing a chromosome.
}
\description{
Bernoulli sampling to form the first generation of the algorithm.
Since interactions are possible, 2*C is a valid population size but
imposing an upper limit of 50 doesn't hurt.
}

