% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initialization}
\alias{initialization}
\title{Generate The 1st Generation of the Genetic Algorithm}
\usage{
initialization(C)
}
\arguments{
\item{C}{chromosome length (Number of Variables in the data)}
}
\value{
Boleans Matrix, [C x 2C] or [C x 50], with each column
representing a chromosome, where T marks that the gene (variable)
as active and F as inactive.
}
\description{
Bernoulli sampling to form the first generation of the algorithm.
Since interactions are possible, 2*C is a valid population size but
imposing an upper limit of 50 doesn't hurt.
}
\details{
This function produces first generation given the chromosome length
for Genetic Algorithm. The row of the generated boolean matrix represents the
locus of a gene(variable), and the column of the matrix represents different
members of the first generation(different models). The number of first
generation is defined by minimum(2C,50).
}
\examples{
initialization(10)
initialization(20)
}

