% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mutation}
\alias{mutation}
\title{Chromosome Mutation}
\usage{
mutation(ind)
}
\arguments{
\item{ind}{a numeric vector represent a individual chromosome.}
}
\value{
It returns a mutated chromosone vector with the same length as input one.
}
\description{
Make a chromosome with fixed probability 0.01 to mute
}
\details{
This function makes a chromosome has a 1% fixed chance to mute in each locale. If mutation
happens at one locale, it will make the value in that locale from 1 to 0(or 0 to 1).
}
\examples{
ind=init(9)[,1] #Generate a chromosome
mutation(ind)
}

