% Generated by roxygen2 (4.0.2): do not edit by hand
\name{produce}
\alias{produce}
\title{Produce Next Generation}
\usage{
produce(pop)
}
\arguments{
\item{pop}{selected parent chromosomes matrix from selection\{GARVaS\}}
}
\value{
It returns next generation matrix with each column represents
new generation chromosome.
}
\description{
Breeding next generation from selected parent chromosomes.
}
\details{
This function produces next generation by randomly choose 2 parent chromosomes from
selected parent matrix produced by selection\{GARVaS\}
to perform the process of crossover and mutation which is based on the crossover\{GARVaS\}
}
\examples{
library(faraway)
dat=ozone #Get ozone data from faraway package
C=dim(dat)[2]-1 #Number of variables
pop=init(C) #produce boleans matrix
pop=selection(pop,f=AIC,dat) #Generare a selected parent matrix
produce(pop) #Produce next generation.
}

