% Generated by roxygen2 (4.0.2): do not edit by hand
\name{production}
\alias{production}
\title{Produce Next Generation}
\usage{
production(pop)
}
\arguments{
\item{pop}{Selected parent chromosome matrix from \code{GARVaS::selection()}}
}
\value{
It returns next generation matrix with each
column representing new generation chromosome.
}
\description{
Breeding next generation from selected parent chromosomes.
}
\details{
This function produces the next generation by randomly choosing
2 parent chromosomes from selected parent matrix produced to perform the
process of crossover and mutation using \code{GARVaS::crossover()}
}
\examples{
library(faraway)
dat<-ozone #Get ozone data from faraway package
C<-dim(dat)[2]-1 #Number of variables
pop<-initialization(C) #produce boleans matrix
production(pop) #Produce next generation.
}

