% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select}
\alias{select}
\title{Genetic Algorithm for Variable Selection}
\usage{
select(dat,generations=10,f=AIC,model="lm")
select(dat,generations=10,f=BIC,model="glm")
}
\arguments{
\item{dat}{data frame containing the predictors in the model.
First column should be the response variable.}

\item{generations}{number of generations users want to generate.
If the f argument is missing, the default is 10.}

\item{f}{function that is defined by users for calculating fitting scores.
Users can choose AIC or BIC or even define by themselves.
If the f argument is missing, the default is AIC.}

\item{model}{the linear model that user wants to use to fit in the data.
It can be either "lm" or "glm", with the default being "lm".}

\item{family}{the family argument in the glm() function,
if model was selected as "glm".}
}
\value{
It returns a list containg the last generation chromosomes matrix,
the fittest chromosomes, the fitting scores of each chromosome in the last
generation and the best model selected.
}
\description{
This function selects best model for users based on genetic algorithm.
}
\details{
This function will give users best linear regression model selected
by genetic algorithm. It allows users to choose when to stop the algorithm
and what kind of criteria they want to implement in the algorithm.
}
\section{Parallel Computation}{
 The internal fitness calculations use parallel
for-loop to rank the chromosomes. By default the parallel back-end will scale
to use all cores on the machine of choice.
}
\examples{
library(faraway)
dat<-ozone #Get ozone data from faraway package
select(dat)
select(dat,generations=25,f=AIC)
}

