% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selection}
\alias{selection}
\title{Select variables using the genetic algorithm}
\usage{
selection(pop, f = AIC, dat)
}
\arguments{
\item{pop}{Population matrix, such as the parent generated using init().}

\item{f}{Fitness function that takes in an lm or glm model and returns a
a numerical 'qualification' of that model. Defaults to AIC.}

\item{dat}{Data frame with columns as predictors.
First column should be the response variable.}
}
\value{
list
}
\description{
Selecting parents based on fitness ranks, with AIC as the default fitness criteria.
Alternatively, we can use tournament selection.
}

