% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test}
\alias{test}
\title{Verify the algorithm on built-in R dataset mtcars}
\usage{
test()
}
\description{
The Motor Trend Car Road Test is a well-established and popular
dataset provided in the base R datasets package. The dataset contains
11 variables with 32 observations. Since the dataset has been extensivelly
analyzed by many statisticians, it provided a good basis for analyzing
the effectiveness of the genetic algorithm variable selection.
}
\details{
For the purposes of this test the first column of mtcars - the MPG rating -
will serve as the response and the other 10 columns will be the predictors.

The basis for comparison will be a manual analysis of the dataset.
}

