% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update}
\alias{update}
\title{Genetic Algorithm for Variable Selection}
\usage{
update(dat,generations=10,f=AIC,graph=TRUE)
update(dat,generations=10,f=BIC,graph=TRUE)
}
\arguments{
\item{dat}{data frame containing the predictors in the model.
First column should be the response variable.}

\item{generations}{number of generations users want to generate.
If the f argument is missing, the default is 10.}

\item{f}{function that is defined by users for calculating fitting scores. Users can choose AIC or BIC or even define by themselvies.
If the f argument is missing, the default is AIC.}

\item{graph}{if graph=TRUE, the function will plot the convergence plot based on fitting score.
If the f argument is missing, the default is TRUE.}
}
\value{
It returns a list containg the last generation chromosomes matrix, the fittest chromosomes,
the fitting scores of each chromosome in the last generation and the best model selected.
}
\description{
This function selects best model for users based on genetic algorithm.
}
\details{
This function will give users best linear regression model selected by genetic algorithm.
It allows users to choose when to stop the algorithm and what kind of criteria they want to implement in
the algorithm. And the convergence graph can also let users know when to stop the algorithm.
}
\examples{
#' library(faraway)
dat=ozone #Get ozone data from faraway package
update(dat,generation=25,f=AIC, graph=TRUE)
}

