% Generated by roxygen2 (4.0.2): do not edit by hand
\name{init}
\alias{init}
\title{Generate the 1st generation of the genetic algorithm}
\usage{
init(C)
}
\arguments{
\item{C}{chromosome length}
}
\value{
gen matrix of the first generation
}
\description{
Generate the 1st generation of the genetic algorithm
}

