% Generated by roxygen2 (4.0.2): do not edit by hand
\name{init}
\alias{init}
\title{Generate The 1st Generation of the Genetic Algorithm}
\usage{
init(C)
}
\arguments{
\item{C}{chromosome length (Number of Variables in the data)}
}
\value{
Boleans Matrix, [C x 2C] or [C x 50], with each column representing a chromosome. , where 1 represent the selecting the gene(variable), while zero not.
}
\description{
Bernoulli sampling to form the first generation of the algorithm.
Since interactions are possible, 2*C is a valid population size but
imposing an upper limit of 50 doesn't hurt.
}
\details{
This function produces first generation given the chromosome length for Genetic Algorithm. The row of
generated booleans matrix represents locus of a gene(variable), and the column of the matrix represents different member
of the first generation(different models). The number of first generation is defined by minimum(2C,50).
}
\examples{
init(C=10)
init(C=20)
}

