% Generated by roxygen2 (4.0.2): do not edit by hand
\name{produce}
\alias{produce}
\title{Produce Next Generation}
\usage{
produce(pop)
}
\arguments{
\item{pop}{Selected parent chromosomes matrix from \code{GARVaS::selection()}}
}
\value{
It returns next generation matrix with each column representing
new generation chromosome.
}
\description{
Breeding next generation from selected parent chromosomes.
}
\details{
This function produces the next generation by randomly choosing 2 parent
chromosomes from selected parent matrix produced by \code{GARVaS::selection()}
to perform the process of crossover and mutation which is based on the \code{GARVaS::crossover()}
}
\examples{
library(faraway)
dat=ozone #Get ozone data from faraway package
C=dim(dat)[2]-1 #Number of variables
pop=init(C) #produce boleans matrix
pop=selection(pop,f=AIC,dat) #Generare a selected parent matrix
produce(pop) #Produce next generation.
}

