% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selection}
\alias{selection}
\title{Select variables using the genetic algorithm}
\title{Select Parents Generated From init\{GARVas\} Based on given Criteria.}
\usage{
selection(pop, f = AIC, dat)
selection(pop,f=AIC,dat)
selection(pop,f=BIC,dat)
}
\arguments{
\item{pop}{Population matrix, such as the parent generated using init().}
\item{pop}{boleans matrix determined by init\{GARVas\}}

\item{f}{Fitness function that takes in an lm or glm model and returns a
a numerical 'qualification' of that model. Defaults to AIC.}
\item{f}{fitness function that takes in an lm or glm model and returns a
a numerical 'qualification' of that model. Users can choose AIC or BIC or even define by themselvies.
If the f argument is missing, the default is AIC.}

\item{dat}{Data frame with columns as predictors.
\item{dat}{data frame containing the predictors in the model.
First column should be the response variable.}
}
\value{
list
Returng a list, where containing chosen parents, fittest parent and fitting score for each parent based on fitting function.
}
\description{
Selecting parents based on fitness ranks, with AIC as the default fitness criteria.
Alternatively, we can use tournament selection.
}
\details{
This function selects parents to breed by fitting linear regression model to each possible parents generated by init\{GARVas\} based on fitness rank
with AIC as the default fitness function.
}
\examples{
library(faraway)
dat=ozone #Get ozone data from faraway package
C=dim(dat)[2]-1 #Number of variables
pop=init(C) #produce boleans matrix
selection(pop,f=AIC,dat)
}

