% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test}
\alias{test}
\title{Test the genetic algorithm implementation on a real dataset}
\usage{
test()
}
\description{
Use the mtcars dataset to model the mpg response to the remaining predictors.
The results of the \code{GARVaS::select()} are compared to manual variable selection.
}

