% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test}
\alias{test}
\title{Manual Testing of Underlying GARVaS Functions}
\usage{
test()
}
\description{
Normally the package is tested automatically by running
\code{devtools::test()} but for simplicity the \code{test()} function
will run all unit tests when sourced and executed.
}
\details{
Functionality is provided by \code{testthat} package so it needs to
be installed in order for the tests to execute.
}

